+++ TypeContractor/TypeScript/ApiClientWriter.cs
@@ -155,7 +155,7 @@ public class ApiClientWriter(string outputPath, string? relativeRoot)
                 }
 
                 var outputType = allTypes.First(x => x.FullName == (returnType.InnerType?.FullName ?? returnType.FullName));
-                var importPath = $"{relativeRoot}/{outputType.ContractedType.Folder.Path.Replace('\\', '/')}/{outputType.ContractedType.Name}";
+                var importPath = $"{relativeRoot}/{outputType.ContractedType.Folder.Path.Replace('\\', '/')}/{outputType.Name}";
                 var returnImport = $"import {{ {string.Join(", ", importTypes)} }} from \"{importPath}\";";
                 imports.Add(returnImport);
             }
@@ -167,7 +167,7 @@ public class ApiClientWriter(string outputPath, string? relativeRoot)
                     continue;
 
                 var outputType = allTypes.First(x => x.FullName == (parameter.Type.InnerType?.FullName ?? parameter.Type.FullName));
-                var importPath = $"{relativeRoot}/{outputType.ContractedType.Folder.Path.Replace('\\', '/')}/{outputType.ContractedType.Name}";
+                var importPath = $"{relativeRoot}/{outputType.ContractedType.Folder.Path.Replace('\\', '/')}/{outputType.Name}";
                 var parameterImport = $"import {{ {parameter.Type.ImportType} }} from '{importPath}'";
                 imports.Add(parameterImport);
             }