+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improve detection of FromRoute/FromQuery parameters in API clients
 - Make sure query parameters are stringified for API clients
-- Fix import paths for nested classes
+- Fix import paths for nested classes for API clients
+- Improve handling of builtin return types for API clients
 
 ## [0.12.0] - 2024-09-13