+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.12.2] - 2024-09-16
+
 ### Fixed
 
 - Handle enumerable parameters in API clients
@@ -228,7 +230,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.12.1...HEAD
+[unreleased]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.12.2...HEAD
+[0.12.2]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.12.2...v0.12.1
 [0.12.1]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.12.1...v0.12.0
 [0.12.0]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.12.0...v0.11.0
 [0.11.0]: https://github.com/PerfectlyNormal/TypeContractor/compare/v0.11.0...v0.10.0